Small Changes Before Cloud

+++ PlanFort/PlanFort/Models/DomainModels/WeatherVM.cs
@@ -8,7 +8,7 @@ namespace PlanFort.Models.DomainModels
     public class WeatherVM
     {
         public string Name { get; set; }
-        public double Temp { get; set; }
+        public int Temp { get; set; }
         public string Description { get; set; }
         public string Icon { get; set; }
     }
